measurements/python_measurement/pyproject.toml
[tool.poetry]
name = "nidcpower_source_dc_voltage"
version = "0.5.0"
package-mode = false
description = "Measurement plug-in example that sources and measures a DC voltage with an NI SMU with a custom logging service."
authors = ["National Instruments"]

[tool.poetry.dependencies]
python = "^3.8"
nidcpower = { version = ">=1.4.4", extras = ["grpc"] }
ni-measurement-plugin-sdk-service = {version = "^2.0.0"}
click = ">=7.1.2, !=8.1.4" # mypy fails with click 8.1.4: https://github.com/pallets/click/issues/2558
grpcio = "*"

[tool.poetry.group.dev.dependencies]
ni-python-styleguide = ">=0.4.1"
mypy = ">=1.0"
grpc-stubs = "^1.53"

[tool.black]
extend_exclude = '\.tox/|_pb2(_grpc)?\.(py|pyi)$'
line-length = 100

[tool.ni-python-styleguide]
extend_exclude = '.tox/,*_pb2_grpc.py,*_pb2_grpc.pyi,*_pb2.py,*_pb2.pyi'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
exclude = ["^stubs/.*"]

[[tool.mypy.overrides]]
module = [
    "hightime.*",
    "nidcpower.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["stubs.*"]
ignore_errors = true
